aghservices_messaging_setup does not need it's second parameter really
Modify the function prototype, code and it's only caller accordingly.
This needs to be done for the core version, too. The wrong decision may
have been taken due to trying to build a single function for the two
different cases (core vs threads).
However, at the end we have two distinct functions, so this doesn't make
sense anymore.