agh_ubus_logstream.c: tweak log message in agh_ubus_logstream_init to be more specific
(we are not talking about the real ubus context, but the AGH one, on which the real context is).
Use logging in agh_ubus_logstream_deinit, and remove a pointless assignment.
Describe function intended behaviour. Update agh_ubus_logstream_init description to reflect the fact the function may terminate the program uncleanly.