agh_mm_handler.c: list and delete messages; list modems
Some object leaks have been experienced with this code, but we where unable to diagnose exactly what's going on.
Infact, at some poin we where not able to reproduce the leaks themselves.
Move messaging init to it's gate: something that probably will need to be done for all the supported functionalities, more or less.