agh_modem.c: set a higher value for timeout
Actually, a crazy one, if you consider a multi-modem scenario. We should also consider the fact that, if we find ourselves in a situation where we should wait so much, maybe something isn't going right.
This change aims to reduce the chance of us leaking MMBearer and/or other objects.
That chance is still not zero at the moment.